% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangrove.modular.asfunction.R
\name{adaptation_comparison}
\alias{adaptation_comparison}
\title{adaptation_comparison}
\usage{
adaptation_comparison(baseline.surge.by.return, return.prob, climate.scenario,
  mangrove.atten, decades, discount, plot.allK = T)
}
\arguments{
\item{return.prob}{return probability for each surge in baseline.surge}

\item{climate.scenario}{an array that gives sea level rise (m) and
change in storm intensity (percent) per decade, multiple scenarios can be input}

\item{mangrove.atten}{vector of mangrove attenuation coeffcients}

\item{decades}{vectors of decade names}

\item{discount}{discount rate}

\item{baseline.surge}{baseline surges (before climate change) (m)}
}
\value{
list with means for noadapt, mangrove, mangrove.allK
}
\description{
compare and plot NPV of no adaptation and mangroves
}

