% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_climatebased_surge.R
\name{compute_climatebased_surge}
\alias{compute_climatebased_surge}
\title{compute climate based surge}
\usage{
compute_climatebased_surge(baseline.surge, climate.scenario, ndecades)
}
\arguments{
\item{baseline.surge}{baseline surges (before climate change) (m)}

\item{climate.scenario}{an array that gives sea level rise (m) and
change in storm intensity (percent) per decade, multiple scenarios can be input}

\item{ndecades}{number of decades}
}
\value{
surge (m)
}
\description{
computes surge for future decades given climate change scenarios
as a function of previous surge and climate change effects on sea 
level rise and storm intensity
}

