% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangrove_adjustment_to_surge.R
\name{mangrove_adjustment_to_surge}
\alias{mangrove_adjustment_to_surge}
\title{mangrove_adjustment_to_surge}
\usage{
mangrove_adjustment_to_surge(surge, attenK)
}
\arguments{
\item{surge}{(m) original surge depth}

\item{attenK}{(percent) attenuation coefficient}
}
\value{
new surge depth (m)
}
\description{
compute the new surge after mangroves
}

