\name{MangroveExamples}
\alias{MangroveExamples}
\alias{ccped}
\alias{contped}
\alias{famped}
\alias{exampleORs}
\alias{exampleBetas}
\docType{data}
\title{
Example data for the Mangrove vignette
}
\description{
Examples of the \code{\link[Mangrove]{MangroveORs}}, \code{\link[Mangrove]{MangroveBetas}} and \code{\link[Mangrove]{MangrovePed}} classes from the \code{\link[Mangrove]{Mangrove}} package. To be used in combination with the Mangrove vignette.
}
\usage{data(ccped)
data(famped)
data(exampleORs)
data(contped)
data(exampleBetas)}
\details{
See vignette("Mangrove") for details on these example objects.
}
\examples{
data(ccped)
summary(ccped)

data(contped)
summary(contped)

data(famped)
summary(famped)

data(exampleORs)
summary(exampleORs)

data(exampleBetas)
summary(exampleBetas)
}
\keyword{datasets}
