\name{MangroveExamples}
\alias{MangroveExamples}
\alias{ccped}
\alias{famped}
\alias{exampleORs}
\docType{data}
\title{
Example data for the Mangrove vignette
}
\description{
Examples of the \code{\link[Mangrove]{MangroveORs}} and \code{\link[Mangrove]{MangrovePed}} classes from the \code{\link[Mangrove]{Mangrove}} package. To be used in combination with the Mangrove vignette.
}
\usage{data(ccped)
data(famped)
data(exampleORs)}
\details{
See vignette("Mangrove") for details on these example objects.
}
\examples{
data(ccped)
summary(ccped)

data(famped)
summary(famped)

data(exampleORs)
summary(exampleORs)
}
\keyword{datasets}
